python/requirements.txt
psutil
mysql-connector-python
python-dotenv
tabulate